Feature/auth

+++ src/main/java/com/getbridge/homework/rest/controller/OneOnOneController.java
@@ -1,5 +1,7 @@
 package com.getbridge.homework.rest.controller;
 
+
+import com.getbridge.homework.rest.config.MyServletRequestWrapper;
 import com.getbridge.homework.rest.dto.Search1on1Dto;
 import com.getbridge.homework.rest.entity.OneOnOne;
 import com.getbridge.homework.rest.dto.OneOnOneDto;
@@ -7,8 +9,12 @@ import com.getbridge.homework.rest.repository.OneOnOneRepository;
 import com.getbridge.homework.rest.service.Service;
 import com.getbridge.homework.rest.service.util.Util;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletRequestWrapper;
 import java.util.List;
 import java.util.Optional;
 
@@ -34,7 +40,11 @@ public class OneOnOneController {
     }
 
     @PostMapping("/create1on1")
-    public ResponseEntity<OneOnOne> createOneOnOne(@RequestBody OneOnOneDto oneOnOne) {
+    public ResponseEntity<OneOnOne> createOneOnOne(@RequestBody OneOnOneDto oneOnOne, HttpServletRequest request) {
+        String userId = (String) request.getAttribute("X-AUTHENTICATED-USER");
+        if (userId == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
         OneOnOne newOneOnOne = oneOnOneRepository.save(util.dtoToOneOnOne(oneOnOne));
         return ResponseEntity.ok(newOneOnOne);
     }
@@ -46,15 +56,32 @@ public class OneOnOneController {
     }
 
     @GetMapping("/get1on1byid/{oneOnOneId}")
-    public ResponseEntity<OneOnOne> getOneOnOneById(@PathVariable String oneOnOneId) {
+    public ResponseEntity<OneOnOne> getOneOnOneById(@PathVariable String oneOnOneId, HttpServletRequestWrapper request) throws IllegalAccessException {
+        String userId = request.getHeader("X-AUTHENTICATED-USER");
+        if (userId == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
         Optional<OneOnOne> oneOnOne = oneOnOneRepository.findById(oneOnOneId);
-        return oneOnOne.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+
+        List<String> ids = oneOnOne.map(OneOnOne::getParticipantIds).orElse(null);
+        if(ids != null && ids.contains(userId)) {
+            return ResponseEntity.ok(oneOnOne.get());
+        } else {
+            throw new IllegalAccessException("user not in 1on1");
+        }
+
     }
 
     @PutMapping("/update1on1/{oneOnOneId}")
-    public ResponseEntity<OneOnOne> updateOneOnOne(@PathVariable String oneOnOneId, @RequestBody OneOnOneDto oneOnOneDto) throws IllegalAccessException {
-        OneOnOne oneOnOne = service.update1on1(oneOnOneId, oneOnOneDto);
+    public ResponseEntity<OneOnOne> updateOneOnOne(@PathVariable String oneOnOneId, @RequestBody OneOnOneDto oneOnOneDto, HttpServletRequestWrapper request) throws IllegalAccessException {
+        String userId = request.getHeader("X-AUTHENTICATED-USER");
+        if (userId == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+        OneOnOne oneOnOne = oneOnOneRepository.findById(oneOnOneId).orElseThrow(() -> new IllegalAccessException("1on1 not found"));
         if(!oneOnOne.isConcluded()) {
+            oneOnOne = service.update1on1(oneOnOneId, oneOnOneDto);
             return oneOnOne.getId() == "empty" ? ResponseEntity.notFound().build() : ResponseEntity.ok(oneOnOne);
         } else {
             throw new IllegalAccessException("1on1 is concluded-readonly");
@@ -62,7 +89,11 @@ public class OneOnOneController {
     }
 
     @DeleteMapping("/deleteoneonone/{oneOnOneId}")
-    public ResponseEntity<Void> deleteOneOnOne(@PathVariable String oneOnOneId) {
+    public ResponseEntity<Void> deleteOneOnOne(@PathVariable String oneOnOneId, MyServletRequestWrapper request) {
+        String userId = request.getHeader("X-AUTHENTICATED-USER");
+        if (userId == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
         Optional<OneOnOne> existingOneOnOne = oneOnOneRepository.findById(oneOnOneId);
         if (existingOneOnOne.isPresent()) {
             oneOnOneRepository.deleteById(oneOnOneId);
@@ -73,16 +104,26 @@ public class OneOnOneController {
     }
 
     @PutMapping("/conclude/{oneOnOneId}")
-    public ResponseEntity<OneOnOne> concludeOneOnOne(@PathVariable String oneOnOneId) {
+    public ResponseEntity<OneOnOne> concludeOneOnOne(@PathVariable String oneOnOneId, HttpServletRequest request) {
+        String userId = request.getHeader("X-AUTHENTICATED-USER");
+        if (userId == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
             return ResponseEntity.ok(service.conclude1on1(oneOnOneId));
     }
 
 
 
     @PostMapping("search/1on1s")
-    public List<OneOnOne> search1on1s (@RequestBody Search1on1Dto search1on1Dto) {
+    public List<OneOnOne> search1on1s (@RequestBody Search1on1Dto search1on1Dto, HttpServletRequest request) {
+        String userId = request.getHeader("X-AUTHENTICATED-USER");
+        if (userId == null) {
+            return null;
+        }
         return service.search(search1on1Dto);
 
     }
+
+
 
 }
